+++ FallRiver/FallRiver/Bullet.cpp
@@ -51,7 +51,7 @@ void Bullet::HandleEvent(Event* pEvent)
 
 RECT Bullet::GetRect()
 {
-	RECT cRect = {GetPosX(), GetPosY(), GetPosX()+GetWidth(), GetPosY()+GetHeight() };
+	RECT cRect = {long(GetPosX()), long(GetPosY()), long(GetPosX()+GetWidth()), long(GetPosY()+GetHeight()) };
 	return cRect;
 }
