+++ FallRiver/FallRiver/Bush.cpp
@@ -32,7 +32,7 @@ void Bush::Render()
 
 RECT Bush::GetRect()
 {
-	RECT cRect = {GetPosX(), GetPosY(), GetPosX()+GetWidth(), GetPosY()+GetHeight() };
+	RECT cRect = {long(GetPosX()), long(GetPosY()), long(GetPosX()+GetWidth()), long(GetPosY()+GetHeight()) };
 	return cRect;
 }
