+++ FallRiver/FallRiver/Closet.cpp
@@ -23,7 +23,7 @@ void Closet::Render()
 
 RECT Closet::GetRect()
 {
-	RECT cRect = {GetPosX(), GetPosY(), GetPosX()+GetWidth(), GetPosY()+GetHeight() };
+	RECT cRect = {long(GetPosX()), long(GetPosY()), long(GetPosX()+GetWidth()), long(GetPosY()+GetHeight()) };
 	return cRect;
 }
