+++ FallRiver/FallRiver/Light.cpp
@@ -25,7 +25,7 @@ void Light::Render()
 
 RECT Light::GetRect()
 {
-	RECT cRect = {GetPosX(), GetPosY(), GetPosX()+GetWidth(), GetPosY()+GetHeight() };
+	RECT cRect = {long(GetPosX()), long(GetPosY()), long(GetPosX()+GetWidth()), long(GetPosY()+GetHeight()) };
 	return cRect;
 }
