+++ FallRiver/FallRiver/Weapon.cpp
@@ -73,7 +73,7 @@ void Weapon::Render()
 
 RECT Weapon::GetRect()
 {
-	RECT cRect = {GetPosX(), GetPosY(), GetPosX()+GetWidth(), GetPosY()+GetHeight() };
+	RECT cRect = {long(GetPosX()), long(GetPosY()), long(GetPosX()+GetWidth()), long(GetPosY()+GetHeight()) };
 	return cRect;
 }
